requirements.txt
sqlalchemy
discord
python-dotenv
requests